+++ test/project/uiengine/pages/patterns/organisms/page.config.js
@@ -1,3 +1,4 @@
 module.exports = {
-  components: ['form']
+  components: ['form'],
+  collapsed: true
 }